backend/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
PyMongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
APScheduler==3.10.4
requests==2.31.0
numpy==1.25.2
pandas==2.1.0
scikit-learn==1.3.0
gunicorn==21.2.0
python-dateutil==2.8.2
email-validator==2.0.0
Pillow==10.0.0
geopy==2.4.0
folium==0.14.0
plotly==5.16.1
eventlet==0.33.3
